+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# NEXT UP
+# 3.4.21
 
 - Counterspell, Feather Fall, Slow Fall, Swiftstride Reaction, and War Caster set to manual reaction type for midi-qol.
 - V11 Support: Support top level folder for compendiums.- V11 Support: First pass of macros.